+++ resharper/resharper-unity/src/CSharp/Feature/Services/CodeCompletion/UnityEventFunctionRule.cs
@@ -18,6 +18,7 @@ using JetBrains.ReSharper.Feature.Services.CodeCompletion.LookupItems.Presentati
 using JetBrains.ReSharper.Feature.Services.CSharp.CodeCompletion.Infrastructure;
 using JetBrains.ReSharper.Feature.Services.Descriptions;
 using JetBrains.ReSharper.Feature.Services.Lookup;
+using JetBrains.ReSharper.Plugins.Unity.ProjectModel;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CSharp;
 using JetBrains.ReSharper.Psi.CSharp.Impl;
