+++ resharper/resharper-unity/src/CSharp/Psi/Resolve/UnityObjectTypeReferenceProviderFactory.cs
@@ -1,5 +1,6 @@
 using JetBrains.DataFlow;
 using JetBrains.Lifetimes;
+using JetBrains.ReSharper.Plugins.Unity.ProjectModel;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.Caches;
 using JetBrains.ReSharper.Psi.CSharp;
