+++ resharper/resharper-unity/src/Settings/LangVersionProjectSettingsProvider.cs
@@ -7,6 +7,7 @@ using JetBrains.Application.Settings.Implementation;
 using JetBrains.Lifetimes;
 using JetBrains.ProjectModel;
 using JetBrains.ProjectModel.Properties.CSharp;
+using JetBrains.ReSharper.Plugins.Unity.ProjectModel;
 using JetBrains.ReSharper.Plugins.Unity.ProjectModel.Caches;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CSharp;
