+++ resharper/resharper-unity/src/VisualStudio/ProjectModel/VsUnityVersionPropertiesExtenderProvider.cs
@@ -7,6 +7,7 @@ using JetBrains.Application.Threading;
 using JetBrains.Lifetimes;
 using JetBrains.ProjectModel;
 using JetBrains.ProjectModel.PropertiesExtender;
+using JetBrains.ReSharper.Plugins.Unity.ProjectModel;
 
 namespace JetBrains.ReSharper.Plugins.Unity.VisualStudio.ProjectModel
 {
